bug(source_spectra): fix number time frames (#155)
* fix number time frames

A bug fix, where start and stop times for each file is used to determine number of time frames to be read. Since time_delta is specified in the Finder, current fix is simply using the start and end times of the files listed by Finder. In future, we may want a more careful selection of frames based on the span of source in a given file.

Co-authored-by: Saurabh Singh <saurabhs24@coconut.chime>